Move opentelemetry tests.
